Updated main.c

+++ GPS_Tracking_System/GPS_interface.h
@@ -1,13 +1,14 @@
+
 #ifndef __HAL_GPS_INTERFACE_H__
 #define __HAL_GPS_INTERFACE_H__
 
-double Calc_Distance(volatile double latitude_1, volatile double longitude_1, volatile double latitude_2, volatile double longitude_2);
+f32 Calc_Distance(volatile f32 latitude_1, volatile f32 longitude_1, volatile f32 latitude_2, volatile f32 longitude_2);
 
 void GPS_Read();
 
 void GPS_format();
 
-double Value_to_Degree(double value);
+f32 Value_to_Degree(f32 value);
 
 #define DEG_TO_RAD(DEG) ((DEG * PI / 180))
 #define M_PI            3.14159265358979323846264338327950
